pyproject.toml
[project]
name = "fluxer-rolebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "emoji>=2.15.0",
    "fluxer-py>=0.3.1",
    "python-dotenv>=1.2.1",
    "tinydb>=4.8.2",
]
